refactor: Added switch to module
