+++ services/export/export.js
@@ -1,15 +1,15 @@
-import { handle } from './handle'
+const { handle } = require('./utils/handle')
 
 module.exports = {
   name: 'export',
   actions: {
     async listCollections (ctx) {
       const { project } = ctx.params
-      handle(project, ctx, 'method')
+      handle(project, ctx)
     },
     async exportCollection (ctx) {
       const { collection, project } = ctx.params
-      handle(project, collection, ctx, 'method')
+      handle(project, collection, ctx)
     }
   }
 }
